+++ pkg/components/cnudie/componentresolvers/blueprint.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package oci
+package componentresolvers
 
 import (
 	"context"
