+++ pkg/components/cnudie/componentresolvers/oci.go
@@ -1,4 +1,4 @@
-package oci
+package componentresolvers
 
 import (
 	"context"
