+++ pkg/components/cnudie/componentresolvers/testdata/comp1/component-descriptor.yaml
