+++ pkg/components/cnudie/componentresolvers/testdata/comp1/config.json
