+++ pkg/landscaper/blueprints/resolve_test.go
@@ -9,8 +9,6 @@ import (
 	"io"
 	"math/rand"
 
-	"github.com/gardener/landscaper/pkg/components/cnudie/oci"
-
 	cdv2 "github.com/gardener/component-spec/bindings-go/apis/v2"
 	"github.com/gardener/component-spec/bindings-go/ctf"
 	"github.com/mandelsoft/vfs/pkg/memoryfs"
@@ -22,6 +20,7 @@ import (
 	lsv1alpha1 "github.com/gardener/landscaper/apis/core/v1alpha1"
 	"github.com/gardener/landscaper/apis/mediatype"
 	"github.com/gardener/landscaper/controller-utils/pkg/logging"
+	"github.com/gardener/landscaper/pkg/components/cnudie/componentresolvers"
 	componentstestutils "github.com/gardener/landscaper/pkg/components/testutils"
 	"github.com/gardener/landscaper/pkg/landscaper/blueprints"
 	"github.com/gardener/landscaper/pkg/landscaper/blueprints/bputils"
@@ -87,7 +86,7 @@ var _ = Describe("Resolve", func() {
 			}).BuildResourceToFs(memFs, "blobs/bp.tar", false)
 			Expect(err).ToNot(HaveOccurred())
 
-			blobResolver := oci.NewLocalFilesystemBlobResolver(memFs)
+			blobResolver := componentresolvers.NewLocalFilesystemBlobResolver(memFs)
 			localFSAccess, err := cdv2.NewUnstructured(cdv2.NewLocalFilesystemBlobAccess("bp.tar", mediatype.BlueprintArtifactsLayerMediaTypeV1))
 			Expect(err).ToNot(HaveOccurred())
 
@@ -136,7 +135,7 @@ var _ = Describe("Resolve", func() {
 				},
 			}).BuildResourceToFs(memFs, "blobs/bp.tar", true)
 			Expect(err).ToNot(HaveOccurred())
-			blobResolver := oci.NewLocalFilesystemBlobResolver(memFs)
+			blobResolver := componentresolvers.NewLocalFilesystemBlobResolver(memFs)
 			localFSAccess, err := cdv2.NewUnstructured(cdv2.NewLocalFilesystemBlobAccess("bp.tar",
 				mediatype.NewBuilder(mediatype.BlueprintArtifactsLayerMediaTypeV1).Compression(mediatype.GZipCompression).String()))
 			Expect(err).ToNot(HaveOccurred())
@@ -186,7 +185,7 @@ var _ = Describe("Resolve", func() {
 				},
 			}).BuildResourceToFs(memFs, "blobs/bp.tar", false)
 			Expect(err).ToNot(HaveOccurred())
-			blobResolver := oci.NewLocalFilesystemBlobResolver(memFs)
+			blobResolver := componentresolvers.NewLocalFilesystemBlobResolver(memFs)
 			localFSAccess, err := cdv2.NewUnstructured(cdv2.NewLocalFilesystemBlobAccess("bp.tar",
 				mediatype.NewBuilder(mediatype.BlueprintArtifactsLayerMediaTypeV1).Compression(mediatype.GZipCompression).String()))
 			Expect(err).ToNot(HaveOccurred())
