+++ pkg/landscaper/installations/imports/imports_suite_test.go
@@ -10,12 +10,11 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/gardener/landscaper/pkg/components/cnudie/oci"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/gardener/landscaper/pkg/components/cnudie/componentresolvers"
 	testutils "github.com/gardener/landscaper/test/utils"
 )
 
@@ -26,7 +25,7 @@ func TestConfig(t *testing.T) {
 
 func createDefaultContextsForNamespace(kubeClient client.Client) {
 	// create default repo for all namespaces
-	repoCtx := oci.NewLocalRepository("../testdata/registry")
+	repoCtx := componentresolvers.NewLocalRepository("../testdata/registry")
 	list, err := os.ReadDir("./testdata/state")
 	Expect(err).To(Succeed())
 	for _, d := range list {
