+++ test/landscaper/deployer_blueprints/deployer_blueprint_test.go
@@ -9,8 +9,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/gardener/landscaper/pkg/components/cnudie/oci"
-
 	cdv2 "github.com/gardener/component-spec/bindings-go/apis/v2"
 	"github.com/mandelsoft/vfs/pkg/osfs"
 	"github.com/mandelsoft/vfs/pkg/projectionfs"
@@ -20,6 +18,7 @@ import (
 	"sigs.k8s.io/yaml"
 
 	lsv1alpha1 "github.com/gardener/landscaper/apis/core/v1alpha1"
+	"github.com/gardener/landscaper/pkg/components/cnudie/componentresolvers"
 	"github.com/gardener/landscaper/pkg/components/model"
 	"github.com/gardener/landscaper/pkg/components/registries"
 	"github.com/gardener/landscaper/pkg/deployer/helm"
@@ -57,7 +56,7 @@ func RenderBlueprint(deployer, componentName, version string) *lsutils.RenderedD
 	registryPath := filepath.Join(projectRoot, ".landscaper", deployer, "example")
 	registryAccess, err := registries.NewFactory().NewLocalRegistryAccess(registryPath)
 	Expect(err).ToNot(HaveOccurred())
-	repository := oci.NewLocalRepository(registryPath)
+	repository := componentresolvers.NewLocalRepository(registryPath)
 	repositoryContext, err := cdv2.NewUnstructured(repository)
 	Expect(err).ToNot(HaveOccurred())
