+++ test/utils/cnudie.go
@@ -7,13 +7,12 @@ package utils
 import (
 	"context"
 
-	"github.com/gardener/landscaper/pkg/components/cnudie/oci"
-
 	cdv2 "github.com/gardener/component-spec/bindings-go/apis/v2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	lsv1alpha1 "github.com/gardener/landscaper/apis/core/v1alpha1"
+	"github.com/gardener/landscaper/pkg/components/cnudie/componentresolvers"
 )
 
 // MakeRepositoryContext creates a new oci registry repository context.
@@ -34,7 +33,7 @@ func ExampleRepositoryContext() *cdv2.UnstructuredTypedObject {
 
 // LocalRepositoryContext returns a new 'local' repository context.
 func LocalRepositoryContext(baseurl string) *cdv2.UnstructuredTypedObject {
-	rctx, _ := cdv2.NewUnstructured(&oci.NewLocalRepository(baseurl).ObjectType)
+	rctx, _ := cdv2.NewUnstructured(&componentresolvers.NewLocalRepository(baseurl).ObjectType)
 	return &rctx
 }
